+++ bin/MorphTest.cpp
@@ -6,7 +6,9 @@
 #include "TFile.h"
 #include "RooRealVar.h"
 #include "RooRealSumPdf.h"
-#include "HiggsAnalysis/CombinedLimit/interface/RooMorphingPdf2.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistFunc.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistErrorPropagator.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistFuncWrapper.h"
 #include "HiggsAnalysis/CombinedLimit/interface/th1fmorph.h"
 #include "TRandom3.h"
 #include "TVector.h"
