+++ bin/MorphTest2.cpp
@@ -7,7 +7,9 @@
 #include "RooRealVar.h"
 #include "RooRealSumPdf.h"
 #include "RooWorkspace.h"
-#include "HiggsAnalysis/CombinedLimit/interface/RooMorphingPdf2.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistFunc.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistErrorPropagator.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistFuncWrapper.h"
 #include "HiggsAnalysis/CombinedLimit/interface/th1fmorph.h"
 #include "TRandom3.h"
 #include "TVector.h"
