+++ interface/CMSHistFuncWrapper.h
@@ -1,0 +1,87 @@
+#ifndef CMSHistFuncWrapper_h
+#define CMSHistFuncWrapper_h
+#include <ostream>
+#include <vector>
+#include <memory>
+#include "RooAbsReal.h"
+#include "RooArgSet.h"
+#include "RooListProxy.h"
+#include "RooRealProxy.h"
+#include "RooArgProxy.h"
+#include "Rtypes.h"
+#include "HiggsAnalysis/CombinedLimit/interface/FastTemplate.h"
+#include "HiggsAnalysis/CombinedLimit/interface/Logging.h"
+#include "HiggsAnalysis/CombinedLimit/interface/SimpleCacheSentry.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistErrorPropagator.h"
+#include "HiggsAnalysis/CombinedLimit/interface/CMSHistFunc.h"
+
+class CMSHistFuncWrapperV;
+
+class CMSHistFuncWrapper : public RooAbsReal {
+ public:
+  CMSHistFuncWrapper();
+
+  CMSHistFuncWrapper(const char* name, const char* title, RooRealVar& x,
+              CMSHistFunc & func, CMSHistErrorPropagator & err, unsigned idx);
+
+  CMSHistFuncWrapper(CMSHistFuncWrapper const& other, const char* name = 0);
+
+  virtual TObject* clone(const char* newname) const {
+    return new CMSHistFuncWrapper(*this, newname);
+  }
+  virtual ~CMSHistFuncWrapper() {}
+
+  Double_t evaluate() const;
+
+  void printMultiline(std::ostream& os, Int_t contents, Bool_t verbose,
+                      TString indent) const;
+
+  Int_t getAnalyticalIntegral(RooArgSet& allVars, RooArgSet& analVars,
+                              const char* rangeName = 0) const;
+
+  Double_t analyticalIntegral(Int_t code, const char* rangeName = 0) const;
+
+  inline void setEvalVerbose(unsigned val) { v = val; };
+
+  void updateCache() const;
+
+  friend class CMSHistFuncWrapperV;
+
+ protected:
+  RooRealProxy x_;
+  RooRealProxy func_;
+  RooArgProxy err_;
+
+  mutable FastHisto cache_;
+  unsigned idx_;
+  mutable SimpleCacheSentry sentry_; //!
+
+ private:
+  ClassDef(CMSHistFuncWrapper, 1)
+  mutable CMSHistFunc const* pfunc_;
+  mutable CMSHistErrorPropagator * perr_;
+  int v;
+  mutable bool initialized_; //! not to be serialized
+
+  void initialize() const;
+};
+
+class CMSHistFuncWrapperV {
+ public:
+  CMSHistFuncWrapperV(const CMSHistFuncWrapper &,
+                              const RooAbsData& data,
+                              bool includeZeroWeights = false);
+  void fill(std::vector<Double_t>& out) const;
+
+ private:
+  const CMSHistFuncWrapper& hpdf_;
+  int begin_, end_, nbins_;
+  struct Block {
+    int index, begin, end;
+    Block(int i, int begin_, int end_) : index(i), begin(begin_), end(end_) {}
+  };
+  std::vector<Block> blocks_;
+  std::vector<int> bins_;
+};
+
+#endif
